requirements.txt
# Core dependencies
pipecat-ai[silero,websockets]>=0.0.27
websockets>=11.0
httpx>=0.25.0
pydantic>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Optional for enhanced features
pyaudio>=0.2.11  # For WhisperLive
prometheus-client>=0.18.0  # For metrics
psutil>=5.9.0  # For system monitoring
